Make comparison of biomass to initial biomass take into account initial wood turnover. Remove foliage from all cohorts that die due to low biomass.

+++ src/ISpeciesPNET.cs
@@ -143,7 +143,7 @@ namespace Landis.Extension.Succession.BiomassPnET
         float LeafOnMinT { get; }
 
         // Initial Biomass
-        float InitBiomass { get; }
+        int InitBiomass { get; }
 
     }
 }
